Adds auth0 back to routers
